requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
cartesia==2.0.17
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
deepgram-sdk==4.7.0
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
faiss-cpu==1.13.0
fastapi==0.123.7
flatbuffers==25.9.23
frozenlist==1.8.0
future==1.0.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
humanfriendly==10.0
idna==3.11
iterators==0.2.0
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-openai==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.12
langsmith==0.4.53
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.10
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nltk==3.9.2
numba==0.61.2
numpy==2.2.6
onnxruntime==1.23.2
openai==2.8.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pillow==11.3.0
pipecat-ai==0.0.96
propcache==0.4.1
protobuf==5.29.5
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
PyJWT==2.10.1
pyloudnorm==0.1.1
pyreadline3==3.5.4
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
resampy==0.4.3
scipy==1.16.3
sniffio==1.3.1
soxr==0.5.0.post1
SQLAlchemy==2.0.44
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
twilio==9.8.8
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uuid_utils==0.12.0
uvicorn==0.38.0
wait_for2==0.4.1
websockets==15.0.1
win32_setctime==1.2.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
